webapp/backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
botocore==1.38.38
certifi==2024.8.30
charset-normalizer==3.3.2
distro==1.9.0
Django==5.1.1
django-cors-headers==4.4.0
django-dotenv==1.4.2
django-ratelimit==4.1.0
django-storages==1.14.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.3.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jmespath==1.0.1
openai==1.88.0
optionaldict==0.1.2
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycryptodome==3.20.0
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pyzbar==0.1.9
requests==2.32.3
rsa==4.9
s3transfer==0.11.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2024.1
urllib3==2.2.3
whitenoise==6.8.2
xmltodict==0.13.0
